backend/requirements.txt
boto3==1.34.0
botocore==1.34.0
pyjwt==2.8.0
cryptography==41.0.7
requests==2.31.0
python-dateutil==2.8.2
validators==0.22.0
pydantic==2.5.0
fastapi==0.104.1
mangum==0.17.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0